Extract Tab from TextEditor as TextEditorTab
