+++ src/main/java/net/imagej/ui/swing/script/TextEditor.java
@@ -726,7 +726,7 @@ public class TextEditor extends JFrame implements ActionListener,
 	 *
 	 * @return The currently selected tab. Never null.
 	 */
-	public Tab getTab() {
+	public TextEditorTab getTab() {
 		final int index = tabbed.getSelectedIndex();
 		if (index < 0) {
 			// should not happen, but safety first.
@@ -737,7 +737,7 @@ public class TextEditor extends JFrame implements ActionListener,
 
 			tabbed.setSelectedIndex(0);
 		}
-		return (Tab) tabbed.getComponentAt(index);
+		return (TextEditorTab) tabbed.getComponentAt(index);
 	}
 
 	/**
@@ -746,8 +746,8 @@ public class TextEditor extends JFrame implements ActionListener,
 	 * @param index the index of the tab.
 	 * @return the {@link Tab} at given index or <code>null</code>.
 	 */
-	public Tab getTab(final int index) {
-		return (Tab) tabbed.getComponentAt(index);
+	public TextEditorTab getTab(final int index) {
+		return (TextEditorTab) tabbed.getComponentAt(index);
 	}
 
 	/**
@@ -1286,278 +1286,6 @@ public class TextEditor extends JFrame implements ActionListener,
 		return false;
 	}
 
-	public class Tab extends JSplitPane {
-
-		protected final EditorPane editorPane = new EditorPane(TextEditor.this);
-		protected final JTextArea screen = new JTextArea();
-		protected final JScrollPane scroll;
-		protected boolean showingErrors;
-		private Executer executer;
-		private final JButton runit, killit, toggleErrors;
-
-		public Tab() {
-			super(JSplitPane.VERTICAL_SPLIT);
-			super.setResizeWeight(350.0 / 430.0);
-
-			screen.setEditable(false);
-			screen.setLineWrap(true);
-			screen.setFont(new Font("Courier", Font.PLAIN, 12));
-
-			final JPanel bottom = new JPanel();
-			bottom.setLayout(new GridBagLayout());
-			final GridBagConstraints bc = new GridBagConstraints();
-
-			bc.gridx = 0;
-			bc.gridy = 0;
-			bc.weightx = 0;
-			bc.weighty = 0;
-			bc.anchor = GridBagConstraints.NORTHWEST;
-			bc.fill = GridBagConstraints.NONE;
-			runit = new JButton("Run");
-			runit.addActionListener(new ActionListener() {
-
-				@Override
-				public void actionPerformed(final ActionEvent ae) {
-					runText();
-				}
-			});
-			bottom.add(runit, bc);
-
-			bc.gridx = 1;
-			killit = new JButton("Kill");
-			killit.setEnabled(false);
-			killit.addActionListener(new ActionListener() {
-
-				@Override
-				public void actionPerformed(final ActionEvent ae) {
-					kill();
-				}
-			});
-			bottom.add(killit, bc);
-
-			bc.gridx = 2;
-			bc.fill = GridBagConstraints.HORIZONTAL;
-			bc.weightx = 1;
-			bottom.add(new JPanel(), bc);
-
-			bc.gridx = 3;
-			bc.fill = GridBagConstraints.NONE;
-			bc.weightx = 0;
-			bc.anchor = GridBagConstraints.NORTHEAST;
-			toggleErrors = new JButton("Show Errors");
-			toggleErrors.addActionListener(new ActionListener() {
-
-				@Override
-				public void actionPerformed(final ActionEvent e) {
-					toggleErrors();
-				}
-			});
-			bottom.add(toggleErrors, bc);
-
-			bc.gridx = 4;
-			bc.fill = GridBagConstraints.NONE;
-			bc.weightx = 0;
-			bc.anchor = GridBagConstraints.NORTHEAST;
-			final JButton clear = new JButton("Clear");
-			clear.addActionListener(new ActionListener() {
-
-				@Override
-				public void actionPerformed(final ActionEvent ae) {
-					if (showingErrors) errorScreen.setText("");
-					else screen.setText("");
-				}
-			});
-			bottom.add(clear, bc);
-
-			bc.gridx = 0;
-			bc.gridy = 1;
-			bc.anchor = GridBagConstraints.NORTHWEST;
-			bc.fill = GridBagConstraints.BOTH;
-			bc.weightx = 1;
-			bc.weighty = 1;
-			bc.gridwidth = 5;
-			screen.setEditable(false);
-			screen.setLineWrap(true);
-			final Font font = new Font("Courier", Font.PLAIN, 12);
-			screen.setFont(font);
-			scroll = new JScrollPane(screen);
-			scroll.setPreferredSize(new Dimension(600, 80));
-			bottom.add(scroll, bc);
-
-			super.setTopComponent(editorPane.wrappedInScrollbars());
-			super.setBottomComponent(bottom);
-		}
-
-		/** Invoke in the context of the event dispatch thread. */
-		private void prepare() {
-			editorPane.setEditable(false);
-			runit.setEnabled(false);
-			killit.setEnabled(true);
-		}
-
-		private void restore() {
-			SwingUtilities.invokeLater(new Runnable() {
-
-				@Override
-				public void run() {
-					editorPane.setEditable(true);
-					runit.setEnabled(true);
-					killit.setEnabled(false);
-					executer = null;
-				}
-			});
-		}
-
-		public void toggleErrors() {
-			showingErrors = !showingErrors;
-			if (showingErrors) {
-				toggleErrors.setText("Show Output");
-				scroll.setViewportView(errorScreen);
-			}
-			else {
-				toggleErrors.setText("Show Errors");
-				scroll.setViewportView(screen);
-			}
-		}
-
-		public void showErrors() {
-			if (!showingErrors) toggleErrors();
-			else if (scroll.getViewport().getView() == null) scroll
-				.setViewportView(errorScreen);
-		}
-
-		public void showOutput() {
-			if (showingErrors) toggleErrors();
-		}
-
-		public JTextArea getScreen() {
-			return showingErrors ? errorScreen : screen;
-		}
-
-		boolean isExecuting() {
-			return null != executer;
-		}
-
-		String getTitle() {
-			return (editorPane.fileChanged() ? "*" : "") + editorPane.getFileName() +
-				(isExecuting() ? " (Running)" : "");
-		}
-
-		/** Invoke in the context of the event dispatch thread. */
-		private void execute(final boolean selectionOnly) throws IOException {
-			prepare();
-			final JTextAreaWriter output =
-				new JTextAreaWriter(this.screen, TextEditor.this.log);
-			final JTextAreaWriter errors = new JTextAreaWriter(errorScreen, log);
-			final File file = getEditorPane().getFile();
-			// Pipe current text into the runScript:
-			final PipedInputStream pi = new PipedInputStream();
-			final PipedOutputStream po = new PipedOutputStream(pi);
-			// The Executer creates a Thread that
-			// does the reading from PipedInputStream
-			this.executer = new TextEditor.Executer(output, errors) {
-
-				@Override
-				public void execute() {
-					try {
-						evalScript(file == null ? getEditorPane().getFileName() : file
-							.getAbsolutePath(), new InputStreamReader(pi), output, errors);
-						output.flush();
-						errors.flush();
-						markCompileEnd();
-					}
-					catch (final Throwable t) {
-						output.flush();
-						errors.flush();
-						if (t instanceof ScriptException && t.getCause() != null &&
-							t.getCause().getClass().getName().endsWith("CompileError"))
-						{
-							errorScreen.append("Compilation failed");
-							showErrors();
-						}
-						else {
-							handleException(t);
-						}
-					}
-					finally {
-						restore();
-					}
-				}
-			};
-			// Write into PipedOutputStream
-			// from another Thread
-			try {
-				final String text;
-				if (selectionOnly) {
-					final String selected = editorPane.getSelectedText();
-					if (selected == null) {
-						error("Selection required!");
-						text = null;
-					}
-					else text = selected + "\n"; // Ensure code blocks are terminated
-				}
-				else {
-					text = editorPane.getText();
-				}
-				new Thread() {
-
-					{
-						setPriority(Thread.NORM_PRIORITY);
-					}
-
-					@Override
-					public void run() {
-						final PrintWriter pw = new PrintWriter(po);
-						pw.write(text);
-						pw.flush(); // will lock and wait in some cases
-						try {
-							po.close();
-						}
-						catch (final Throwable tt) {
-							tt.printStackTrace();
-						}
-						pw.close();
-					}
-				}.start();
-			}
-			catch (final Throwable t) {
-				t.printStackTrace();
-			}
-			finally {
-				// Re-enable when all text to send has been sent
-				editorPane.setEditable(true);
-			}
-		}
-
-		protected void kill() {
-			if (null == executer) return;
-			// Graceful attempt:
-			executer.interrupt();
-			// Give it 3 seconds. Then, stop it.
-			final long now = System.currentTimeMillis();
-			new Thread() {
-
-				{
-					setPriority(Thread.NORM_PRIORITY);
-				}
-
-				@Override
-				public void run() {
-					while (System.currentTimeMillis() - now < 3000)
-						try {
-							Thread.sleep(100);
-						}
-						catch (final InterruptedException e) {
-							/* ignore */
-						}
-					if (null != executer) executer.obliterate();
-					restore();
-
-				}
-			}.start();
-		}
-	}
-
 	public static boolean isBinary(final File file) {
 		if (file == null) return false;
 		// heuristic: read the first up to 8000 bytes, and say that it is binary if
@@ -1588,9 +1316,9 @@ public class TextEditor extends JFrame implements ActionListener,
 	 * Open a new tab with some content; the languageExtension is like ".java",
 	 * ".py", etc.
 	 */
-	public Tab newTab(final String content, final String language) {
+	public TextEditorTab newTab(final String content, final String language) {
 		String lang = language;
-		final Tab tab = open(null);
+		final TextEditorTab tab = open(null);
 		if (null != lang && lang.length() > 0) {
 			lang = lang.trim().toLowerCase();
 
@@ -1608,7 +1336,7 @@ public class TextEditor extends JFrame implements ActionListener,
 		return tab;
 	}
 
-	public Tab open(final File file) {
+	public TextEditorTab open(final File file) {
 		if (isBinary(file)) {
 			// TODO!
 			throw new RuntimeException("TODO: open image using IJ2");
@@ -1616,10 +1344,10 @@ public class TextEditor extends JFrame implements ActionListener,
 		}
 
 		try {
-			Tab tab = (tabbed.getTabCount() == 0) ? null : getTab();
+			TextEditorTab tab = (tabbed.getTabCount() == 0) ? null : getTab();
 			final boolean wasNew = tab != null && tab.editorPane.isNew();
 			if (!wasNew) {
-				tab = new Tab();
+				tab = new TextEditorTab(this);
 				context.inject(tab.editorPane);
 				tab.editorPane.loadPreferences();
 				addDefaultAccelerators(tab.editorPane);
@@ -1923,7 +1651,7 @@ public class TextEditor extends JFrame implements ActionListener,
 	}
 
 	synchronized void setTitle() {
-		final Tab tab = getTab();
+		final TextEditorTab tab = getTab();
 
 		final boolean fileChanged = tab.editorPane.fileChanged();
 		final String fileName = tab.editorPane.getFileName();
@@ -1937,7 +1665,7 @@ public class TextEditor extends JFrame implements ActionListener,
 				setTitle(title); // to the main window
 				// Update all tabs: could have changed
 				for (int i = 0; i < tabbed.getTabCount(); i++)
-					tabbed.setTitleAt(i, ((Tab) tabbed.getComponentAt(i)).getTitle());
+					tabbed.setTitleAt(i, ((TextEditorTab) tabbed.getComponentAt(i)).getTitle());
 			}
 		});
 	}
@@ -2099,8 +1827,8 @@ public class TextEditor extends JFrame implements ActionListener,
 
 	public void kill(final Executer executer) {
 		for (int i = 0; i < tabbed.getTabCount(); i++) {
-			final Tab tab = (Tab) tabbed.getComponentAt(i);
-			if (executer == tab.executer) {
+			final TextEditorTab tab = (TextEditorTab) tabbed.getComponentAt(i);
+			if (executer == tab.getExecuter()) {
 				tab.kill();
 				break;
 			}
@@ -2142,12 +1870,101 @@ public class TextEditor extends JFrame implements ActionListener,
 		markCompileStart();
 
 		try {
-			final Tab tab = getTab();
+			final TextEditorTab tab = getTab();
 			tab.showOutput();
-			tab.execute(selectionOnly);
+			execute(selectionOnly);
 		}
 		catch (final Throwable t) {
 			t.printStackTrace();
+		}
+	}
+	
+	/** Invoke in the context of the event dispatch thread. */
+	private void execute(final boolean selectionOnly) throws IOException {
+		final TextEditorTab tab = getTab();
+		
+		tab.prepare();
+		
+		final JTextAreaWriter output =
+			new JTextAreaWriter(tab.screen, log);
+		final JTextAreaWriter errors = new JTextAreaWriter(errorScreen, log);
+		final File file = getEditorPane().getFile();
+		// Pipe current text into the runScript:
+		final PipedInputStream pi = new PipedInputStream();
+		final PipedOutputStream po = new PipedOutputStream(pi);
+		// The Executer creates a Thread that
+		// does the reading from PipedInputStream
+		tab.setExecutor(new Executer(output, errors) {
+
+			@Override
+			public void execute() {
+				try {
+					evalScript(file == null ? getEditorPane().getFileName() : file
+						.getAbsolutePath(), new InputStreamReader(pi), output, errors);
+					output.flush();
+					errors.flush();
+					markCompileEnd();
+				}
+				catch (final Throwable t) {
+					output.flush();
+					errors.flush();
+					if (t instanceof ScriptException && t.getCause() != null &&
+						t.getCause().getClass().getName().endsWith("CompileError"))
+					{
+						errorScreen.append("Compilation failed");
+						tab.showErrors();
+					}
+					else {
+						handleException(t);
+					}
+				}
+				finally {
+					tab.restore();
+				}
+			}
+		});
+		// Write into PipedOutputStream
+		// from another Thread
+		try {
+			final String text;
+			if (selectionOnly) {
+				final String selected = tab.getEditorPane().getSelectedText();
+				if (selected == null) {
+					error("Selection required!");
+					text = null;
+				}
+				else text = selected + "\n"; // Ensure code blocks are terminated
+			}
+			else {
+				text = tab.getEditorPane().getText();
+			}
+			new Thread() {
+
+				{
+					setPriority(Thread.NORM_PRIORITY);
+				}
+
+				@Override
+				public void run() {
+					final PrintWriter pw = new PrintWriter(po);
+					pw.write(text);
+					pw.flush(); // will lock and wait in some cases
+					try {
+						po.close();
+					}
+					catch (final Throwable tt) {
+						tt.printStackTrace();
+					}
+					pw.close();
+				}
+			}.start();
+		}
+		catch (final Throwable t) {
+			t.printStackTrace();
+		}
+		finally {
+			// Re-enable when all text to send has been sent
+			tab.getEditorPane().setEditable(true);
 		}
 	}
 
@@ -2417,13 +2234,13 @@ public class TextEditor extends JFrame implements ActionListener,
 	 * @param message The text to write
 	 */
 	public void write(String message) {
-		final Tab tab = getTab();
+		final TextEditorTab tab = getTab();
 		if (!message.endsWith("\n")) message += "\n";
 		tab.screen.insert(message, tab.screen.getDocument().getLength());
 	}
 
 	public void writeError(String message) {
-		final Tab tab = getTab();
+		final TextEditorTab tab = getTab();
 		tab.showErrors();
 		if (!message.endsWith("\n")) message += "\n";
 		errorScreen.insert(message, errorScreen.getDocument().getLength());
